Simplify FilterExecTransformerBase
