+++ backends-clickhouse/src/main/scala/io/glutenproject/execution/CHFilterExecTransformer.scala
@@ -16,62 +16,12 @@
  */
 package io.glutenproject.execution
 
-import io.glutenproject.extension.ValidationResult
-import io.glutenproject.substrait.SubstraitContext
-
-import org.apache.spark.sql.catalyst.expressions.{And, Expression, PredicateHelper}
+import org.apache.spark.sql.catalyst.expressions.{And, Expression}
 import org.apache.spark.sql.execution.SparkPlan
 
 case class CHFilterExecTransformer(condition: Expression, child: SparkPlan)
-  extends FilterExecTransformerBase(condition, child)
-  with PredicateHelper {
-
-  override protected def doValidateInternal(): ValidationResult = {
-    val remainingCondition = getRemainingCondition
-    if (remainingCondition == null) {
-      // All the filters can be pushed down and the computing of this Filter
-      // is not needed.
-      return ValidationResult.ok
-    }
-    val substraitContext = new SubstraitContext
-    val operatorId = substraitContext.nextOperatorId(this.nodeName)
-    // Firstly, need to check if the Substrait plan for this operator can be successfully generated.
-    val relNode =
-      getRelNode(
-        substraitContext,
-        remainingCondition,
-        child.output,
-        operatorId,
-        null,
-        validation = true)
-    doNativeValidation(substraitContext, relNode)
-  }
-
-  override def doTransform(context: SubstraitContext): TransformContext = {
-    val childCtx = child.asInstanceOf[TransformSupport].doTransform(context)
-    val remainingCondition = getRemainingCondition
-
-    val operatorId = context.nextOperatorId(this.nodeName)
-    if (remainingCondition == null) {
-      // The computing for this filter is not needed.
-      context.registerEmptyRelToOperator(operatorId)
-      // Since some columns' nullability will be removed after this filter, we need to update the
-      // outputAttributes of child context.
-      TransformContext(childCtx.inputAttributes, output, childCtx.root)
-    }
-
-    val currRel = getRelNode(
-      context,
-      remainingCondition,
-      child.output,
-      operatorId,
-      childCtx.root,
-      validation = false)
-    assert(currRel != null, "Filter rel should be valid.")
-    TransformContext(childCtx.outputAttributes, output, currRel)
-  }
-
-  private def getRemainingCondition: Expression = {
+  extends FilterExecTransformerBase(condition, child) {
+  override protected def getRemainingCondition: Expression = {
     val scanFilters = child match {
       // Get the filters including the manually pushed down ones.
       case basicScanTransformer: BasicScanExecTransformer =>
@@ -94,8 +44,8 @@ case class CHFilterExecTransformer(condition: Expression, child: SparkPlan)
 }
 
 case class FilterExecTransformer(condition: Expression, child: SparkPlan)
-  extends FilterExecTransformerBase(condition, child)
-  with TransformSupport {
+  extends FilterExecTransformerBase(condition, child) {
+  override protected def getRemainingCondition: Expression = condition
   override protected def withNewChildInternal(newChild: SparkPlan): FilterExecTransformer =
     copy(child = newChild)
 }
