+++ backends-velox/src/main/scala/io/glutenproject/execution/FilterExecTransformer.scala
@@ -16,61 +16,12 @@
  */
 package io.glutenproject.execution
 
-import io.glutenproject.extension.ValidationResult
-import io.glutenproject.substrait.SubstraitContext
-
-import org.apache.spark.sql.catalyst.expressions.{And, Expression, PredicateHelper}
+import org.apache.spark.sql.catalyst.expressions.{And, Expression}
 import org.apache.spark.sql.execution.SparkPlan
 
 case class FilterExecTransformer(condition: Expression, child: SparkPlan)
-  extends FilterExecTransformerBase(condition, child)
-  with PredicateHelper {
-
-  override protected def doValidateInternal(): ValidationResult = {
-    val remainingCondition = getRemainingCondition
-    if (remainingCondition == null) {
-      // All the filters can be pushed down and the computing of this Filter
-      // is not needed.
-      return ValidationResult.ok
-    }
-    val substraitContext = new SubstraitContext
-    val operatorId = substraitContext.nextOperatorId(this.nodeName)
-    // Firstly, need to check if the Substrait plan for this operator can be successfully generated.
-    val relNode =
-      getRelNode(
-        substraitContext,
-        remainingCondition,
-        child.output,
-        operatorId,
-        null,
-        validation = true)
-    // Then, validate the generated plan in native engine.
-    doNativeValidation(substraitContext, relNode)
-  }
-
-  override def doTransform(context: SubstraitContext): TransformContext = {
-    val childCtx = child.asInstanceOf[TransformSupport].doTransform(context)
-    val remainingCondition = getRemainingCondition
-
-    val operatorId = context.nextOperatorId(this.nodeName)
-    if (remainingCondition == null) {
-      // The computing for this filter is not needed.
-      context.registerEmptyRelToOperator(operatorId)
-      return childCtx
-    }
-
-    val currRel = getRelNode(
-      context,
-      remainingCondition,
-      child.output,
-      operatorId,
-      childCtx.root,
-      validation = false)
-    assert(currRel != null, "Filter rel should be valid.")
-    TransformContext(childCtx.outputAttributes, output, currRel)
-  }
-
-  private def getRemainingCondition: Expression = {
+  extends FilterExecTransformerBase(condition, child) {
+  override protected def getRemainingCondition: Expression = {
     val scanFilters = child match {
       // Get the filters including the manually pushed down ones.
       case basicScanExecTransformer: BasicScanExecTransformer =>
